+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_assets/_artifacts/_package/model_package.py
@@ -77,6 +77,7 @@ class PackageInputPathVersion:
 
     def __init__(
         self,
+        *,
         input_path_type: Optional[str] = None,
         resource_name: Optional[str] = None,
         resource_version: Optional[str] = None,
@@ -114,7 +115,7 @@ class PackageInputPathUrl:
     :type url: str
     """
 
-    def __init__(self, input_path_type: Optional[str] = None, url: Optional[str] = None):
+    def __init__(self, *, input_path_type: Optional[str] = None, url: Optional[str] = None):
         self.input_path_type = input_path_type
         self.url = url
 
@@ -149,6 +150,7 @@ class ModelPackageInput:
 
     def __init__(
         self,
+        *,
         type: Optional[str] = None,
         path: Optional[Union[PackageInputPathId, PackageInputPathUrl, PackageInputPathVersion]] = None,
         mode: Optional[str] = None,