+++ sdk/ml/azure-ai-ml/azure/ai/ml/operations/_data_operations.py
@@ -24,8 +24,8 @@ from azure.ai.ml._artifacts._constants import (
     CHANGED_ASSET_PATH_MSG_NO_PERSONAL_DATA,
 )
 from azure.ai.ml._exception_helper import log_and_raise_error
-from azure.ai.ml._restclient.v2022_10_01_preview.models import ListViewType
-from azure.ai.ml._restclient.v2022_10_01 import AzureMachineLearningWorkspaces as ServiceClient102022
+from azure.ai.ml._restclient.v2023_04_01_preview.models import ListViewType
+from azure.ai.ml._restclient.v2023_04_01_preview import AzureMachineLearningWorkspaces as ServiceClient042023_preview
 from azure.ai.ml._scope_dependent_operations import OperationConfig, OperationScope, _ScopeDependentOperations
 
 from azure.ai.ml._restclient.v2021_10_01_dataplanepreview import (
@@ -38,6 +38,9 @@ from azure.ai.ml._utils._asset_utils import (
     _create_or_update_autoincrement,
     _get_latest_version_from_container,
     _resolve_label_to_asset,
+    _check_or_modify_auto_delete_setting,
+    _validate_auto_delete_setting_in_data_output,
+    _validate_workspace_managed_datastore,
 )
 from azure.ai.ml._utils._data_utils import (
     download_mltable_metadata_schema,
@@ -82,7 +85,7 @@ class DataOperations(_ScopeDependentOperations):
         self,
         operation_scope: OperationScope,
         operation_config: OperationConfig,
-        service_client: Union[ServiceClient102022, ServiceClient102021Dataplane],
+        service_client: Union[ServiceClient042023_preview, ServiceClient102021Dataplane],
         datastore_operations: DatastoreOperations,
         **kwargs: Dict,
     ):
@@ -311,6 +314,9 @@ class DataOperations(_ScopeDependentOperations):
                 artifact_type=ErrorTarget.DATA,
                 show_progress=self._show_progress,
             )
+
+            _check_or_modify_auto_delete_setting(data.auto_delete_setting)
+
             data_version_resource = data._to_rest_object()
             auto_increment_version = data._auto_increment_version
 
@@ -374,6 +380,13 @@ class DataOperations(_ScopeDependentOperations):
 
         experiment_name = "data_import_" + data_import.name
         data_import.type = AssetTypes.MLTABLE if isinstance(data_import.source, Database) else AssetTypes.URI_FOLDER
+
+        # avoid specifying auto_delete_setting in job output now
+        _validate_auto_delete_setting_in_data_output(data_import.auto_delete_setting)
+
+        # block cumtomer specified path on managed datastore
+        data_import.path = _validate_workspace_managed_datastore(data_import.path)
+
         if "${{name}}" not in data_import.path:
             data_import.path = data_import.path.rstrip("/") + "/${{name}}"
         import_job = import_data_func(