+++ sdk/ml/azure-ai-ml/azure/ai/ml/operations/_workspace_operations_base.py
@@ -255,7 +255,7 @@ class WorkspaceOperationsBase:
         return poller
 
     def begin_delete(
-        self, name: str, *, delete_dependent_resources: bool, force_to_purge: bool = False, **kwargs: Dict
+        self, name: str, *, delete_dependent_resources: bool, permanently_delete: bool = False, **kwargs: Dict
     ) -> LROPoller[None]:
         workspace = self.get(name, **kwargs)
         resource_group = kwargs.get("resource_group") or self._resource_group_name
@@ -287,7 +287,7 @@ class WorkspaceOperationsBase:
         poller = self._operation.begin_delete(
             resource_group_name=resource_group,
             workspace_name=name,
-            force_to_purge=force_to_purge,
+            force_to_purge=permanently_delete,
             **self._init_kwargs,
         )
         module_logger.info("Delete request initiated for workspace: %s\n", name)
@@ -463,6 +463,8 @@ class WorkspaceOperationsBase:
         else:
             managed_network = ManagedNetwork(IsolationMode.DISABLED)._to_rest_object()
         _set_val(param["managedNetwork"], managed_network)
+        if workspace.enable_data_isolation:
+            _set_val(param["enable_data_isolation"], "true")
 
         resources_being_deployed[workspace.name] = (ArmConstants.WORKSPACE, None)
         return template, param, resources_being_deployed